requirements.txt
pyserial==3.5
python-dotenv=1.1.1
requests==2.32.5